Add test for CacheRetainingDataFromPreviousBuild
